requirements.txt
amqp==5.0.1
anyjson==0.3.3
APScheduler==3.6.3
asgiref==3.2.10
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==5.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==2.1
django-rest-swagger==2.2.0
djangorestframework==3.11.0
drf-extensions==0.6.0
idna==2.10
importlib-metadata==2.0.0
itypes==1.2.0
Jinja2==2.11.2
kombu==5.0.2
lxml==4.5.2
MarkupSafe==1.1.1
MechanicalSoup==1.0.0.dev0
mysql==0.0.2
mysqlclient==1.4.2.post1
openapi-codec==1.3.2
pkg-resources==0.0.0
prompt-toolkit==3.0.8
PyMySQL==0.10.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
zipp==3.3.1
